+++ lib/Screens/Classes.dart
@@ -5,3 +5,10 @@ class UserClass{
   List<dynamic> markets;
   UserClass({this.name,this.markets,this.userType});
 }
+
+class ProductClass{
+  String name;
+  int price;
+  String additionalInfo;
+  ProductClass({this.name,this.additionalInfo,this.price});
+}
